Gets the the statistics of the first Digits of a benford object

It gets the statistics of the first digits (Frequencies, Squared Differences, Absolute Differences etc). See the section value of benford.

getBfd(bfd)

Arguments

bfd
an object of class "Benford". See benford.

Value

A data.frame with first digits and their statistics.

Examples

data(corporate.payment)
cp <- benford(corporate.payment$Amount) #generates benford object
getBfd(cp) # equivalent to cp$bfd

#>     digits   data.dist data.second.order.dist benford.dist
#>  1:     10 0.056585424           0.3747862726  0.041392685
#>  2:     11 0.055791186           0.0159226917  0.037788561
#>  3:     12 0.032369261           0.0146097948  0.034762106
#>  4:     13 0.031164397           0.0132663654  0.032184683
#>  5:     14 0.024324222           0.0111138251  0.029963223
#>  6:     15 0.030386367           0.0115107474  0.028028724
#>  7:     16 0.023854163           0.0103657792  0.026328939
#>  8:     17 0.021795627           0.0091292135  0.024823584
#>  9:     18 0.020850105           0.0093582071  0.023481096
#> 10:     19 0.020434076           0.0081063752  0.022276395
#> 11:     20 0.021460642           0.1335185638  0.021189299
#> 12:     21 0.019202196           0.0066713483  0.020203386
#> 13:     22 0.016489899           0.0059691011  0.019305155
#> 14:     23 0.015225601           0.0062133610  0.018483406
#> 15:     24 0.014479990           0.0053126527  0.017728767
#> 16:     25 0.020704225           0.0058927699  0.017033339
#> 17:     26 0.015344467           0.0056179775  0.016390416
#> 18:     27 0.012356618           0.0049920616  0.015794267
#> 19:     28 0.013366976           0.0074041280  0.015239967
#> 20:     29 0.012470081           0.0019540791  0.014723257
#> 21:     30 0.017327361           0.0703773815  0.014240439
#> 22:     31 0.012140499           0.0038623596  0.013788284
#> 23:     32 0.011086918           0.0045493405  0.013363962
#> 24:     33 0.010854590           0.0035723009  0.012964977
#> 25:     34 0.009487635           0.0036944309  0.012589127
#> 26:     35 0.011702858           0.0034501710  0.012234456
#> 27:     36 0.010163008           0.0035723009  0.011899223
#> 28:     37 0.009968501           0.0029616512  0.011581873
#> 29:     38 0.008909516           0.0026715926  0.011281010
#> 30:     39 0.008504293           0.0027326575  0.010995384
#> 31:     40 0.010887007           0.0481039326  0.010723865
#> 32:     41 0.006905010           0.0026257938  0.010465434
#> 33:     42 0.010152202           0.0024120664  0.010219165
#> 34:     43 0.007045488           0.0024731314  0.009984221
#> 35:     44 0.007866741           0.0025341964  0.009759837
#> 36:     45 0.010789754           0.0024273327  0.009545318
#> 37:     46 0.007488532           0.0023052027  0.009340026
#> 38:     47 0.007504741           0.0019846116  0.009143379
#> 39:     48 0.007380473           0.0020762091  0.008954843
#> 40:     49 0.006845577           0.0023052027  0.008773924
#> 41:     50 0.040684450           0.0373259648  0.008600172
#> 42:     51 0.006326891           0.0018777479  0.008433168
#> 43:     52 0.007629010           0.0019540791  0.008272526
#> 44:     53 0.007364264           0.0015266243  0.008117890
#> 45:     54 0.006640264           0.0019235467  0.007968930
#> 46:     55 0.006521399           0.0018014167  0.007825338
#> 47:     56 0.006996861           0.0028395213  0.007686829
#> 48:     57 0.005786593           0.0017098192  0.007553138
#> 49:     58 0.006175608           0.0002289936  0.007424018
#> 50:     59 0.007504741           0.0014350269  0.007299239
#> 51:     60 0.008817666           0.0252503664  0.007178585
#> 52:     61 0.005316534           0.0015418906  0.007061854
#> 53:     62 0.006380921           0.0012976307  0.006948860
#> 54:     63 0.005878444           0.0012365657  0.006839425
#> 55:     64 0.007099518           0.0013586957  0.006733383
#> 56:     65 0.006099966           0.0014960918  0.006630579
#> 57:     66 0.005100414           0.0012823644  0.006530867
#> 58:     67 0.005700145           0.0012518319  0.006434110
#> 59:     68 0.004965340           0.0011602345  0.006340178
#> 60:     69 0.004922116           0.0011144358  0.006248949
#> 61:     70 0.005846026           0.0190064729  0.006160309
#> 62:     71 0.004830265           0.0012670982  0.006074148
#> 63:     72 0.005338146           0.0015266243  0.005990364
#> 64:     73 0.003981997           0.0009770396  0.005908860
#> 65:     74 0.005348952           0.0010228383  0.005829544
#> 66:     75 0.007045488           0.0012670982  0.005752329
#> 67:     76 0.004743818           0.0010839033  0.005677133
#> 68:     77 0.004122475           0.0012976307  0.005603878
#> 69:     78 0.004133281           0.0011602345  0.005532489
#> 70:     79 0.004430445           0.0009770396  0.005462896
#> 71:     80 0.006191817           0.0164570103  0.005395032
#> 72:     81 0.004225131           0.0012060332  0.005328834
#> 73:     82 0.004878892           0.0011297020  0.005264240
#> 74:     83 0.004430445           0.0012212995  0.005201194
#> 75:     84 0.004695191           0.0009923058  0.005139640
#> 76:     85 0.005289519           0.0013739619  0.005079526
#> 77:     86 0.006164802           0.0010381045  0.005020801
#> 78:     87 0.003803699           0.0010228383  0.004963420
#> 79:     88 0.004581728           0.0010075721  0.004907334
#> 80:     89 0.006105369           0.0008396434  0.004852503
#> 81:     90 0.008779845           0.0141670738  0.004798883
#> 82:     91 0.004592534           0.0010075721  0.004746435
#> 83:     92 0.008525905           0.0006717147  0.004695121
#> 84:     93 0.005856832           0.0009465071  0.004644905
#> 85:     94 0.003992803           0.0010991695  0.004595752
#> 86:     95 0.006829368           0.0007327797  0.004547628
#> 87:     96 0.006926622           0.0007327797  0.004500501
#> 88:     97 0.003522744           0.0010991695  0.004454341
#> 89:     98 0.009217486           0.0009007084  0.004409119
#> 90:     99 0.007991009           0.0007480459  0.004364805
#>     digits   data.dist data.second.order.dist benford.dist
#>     data.second.order.dist.freq data.dist.freq benford.dist.freq
#>  1:                       24550          10473         7661.0823
#>  2:                        1043          10326         6994.0202
#>  3:                         957           5991         6433.8749
#>  4:                         869           5768         5956.8378
#>  5:                         728           4502         5545.6833
#>  6:                         754           5624         5187.6403
#>  7:                         679           4415         4873.0390
#>  8:                         598           4034         4594.4233
#>  9:                         613           3859         4345.9517
#> 10:                         531           3782         4122.9820
#> 11:                        8746           3972         3921.7790
#> 12:                         437           3554         3739.3033
#> 13:                         391           3052         3573.0560
#> 14:                         407           2818         3420.9642
#> 15:                         348           2680         3281.2934
#> 16:                         386           3832         3152.5815
#> 17:                         368           2840         3033.5874
#> 18:                         327           2287         2923.2504
#> 19:                         485           2474         2820.6587
#> 20:                         128           2308         2725.0245
#> 21:                        4610           3207         2635.6632
#> 22:                         253           2247         2551.9771
#> 23:                         298           2052         2473.4421
#> 24:                         234           2009         2399.5969
#> 25:                         242           1756         2330.0334
#> 26:                         226           2166         2264.3899
#> 27:                         234           1881         2202.3439
#> 28:                         194           1845         2143.6077
#> 29:                         175           1649         2087.9232
#> 30:                         179           1574         2035.0587
#> 31:                        3151           2015         1984.8052
#> 32:                         172           1278         1936.9739
#> 33:                         158           1879         1891.3937
#> 34:                         162           1304         1847.9096
#> 35:                         166           1456         1806.3800
#> 36:                         159           1997         1766.6761
#> 37:                         151           1386         1728.6801
#> 38:                         130           1389         1692.2841
#> 39:                         136           1366         1657.3891
#> 40:                         151           1267         1623.9042
#> 41:                        2445           7530         1591.7456
#> 42:                         123           1171         1560.8359
#> 43:                         128           1412         1531.1039
#> 44:                         100           1363         1502.4835
#> 45:                         126           1229         1474.9134
#> 46:                         118           1207         1448.3369
#> 47:                         186           1295         1422.7013
#> 48:                         112           1071         1397.9574
#> 49:                          15           1143         1374.0595
#> 50:                          94           1389         1350.9650
#> 51:                        1654           1632         1328.6340
#> 52:                         101            984         1307.0292
#> 53:                          85           1181         1286.1158
#> 54:                          81           1088         1265.8612
#> 55:                          89           1314         1246.2347
#> 56:                          98           1129         1227.2074
#> 57:                          84            944         1208.7525
#> 58:                          82           1055         1190.8444
#> 59:                          76            919         1173.4592
#> 60:                          73            911         1156.5743
#> 61:                        1245           1082         1140.1684
#> 62:                          83            894         1124.2215
#> 63:                         100            988         1108.7145
#> 64:                          64            737         1093.6295
#> 65:                          67            990         1078.9494
#> 66:                          83           1304         1064.6583
#> 67:                          71            878         1050.7408
#> 68:                          85            763         1037.1825
#> 69:                          76            765         1023.9696
#> 70:                          64            820         1011.0891
#> 71:                        1078           1146          998.5287
#> 72:                          79            782          986.2765
#> 73:                          74            903          974.3213
#> 74:                          80            820          962.6525
#> 75:                          65            869          951.2599
#> 76:                          90            979          940.1338
#> 77:                          68           1141          929.2650
#> 78:                          67            704          918.6446
#> 79:                          66            848          908.2642
#> 80:                          55           1130          898.1158
#> 81:                         928           1625          888.1916
#> 82:                          66            850          878.4844
#> 83:                          44           1578          868.9871
#> 84:                          62           1084          859.6930
#> 85:                          72            739          850.5955
#> 86:                          48           1264          841.6886
#> 87:                          48           1282          832.9663
#> 88:                          72            652          824.4229
#> 89:                          59           1706          816.0530
#> 90:                          49           1479          807.8513
#>     data.second.order.dist.freq data.dist.freq benford.dist.freq
#>     benford.so.dist.freq data.summation abs.excess.summation  difference
#>  1:            2711.3864     28701407.0          23224143.17  2811.91765
#>  2:            2475.3019     22324747.7          16847483.89  3331.97978
#>  3:            2277.0570     16258127.0          10780863.16  -442.87491
#>  4:            2108.2255     15520164.5          10042900.72  -188.83775
#>  5:            1962.7110     27393259.5          21915995.69 -1043.68327
#>  6:            1835.9935     49191987.5          43714723.72   436.35975
#>  7:            1724.6508     12523174.3           7045910.54  -458.03897
#>  8:            1626.0440     11994778.3           6517514.50  -560.42335
#>  9:            1538.1057      7545938.7           2068674.89  -486.95166
#> 10:            1459.1930      6987397.0           1510133.24  -340.98196
#> 11:            1387.9838     12555177.0           7077913.19    50.22096
#> 12:            1323.4026      6145102.1            667838.28  -185.30331
#> 13:            1264.5649      7134018.0           1656754.17  -521.05604
#> 14:            1210.7370      6779017.1           1301753.32  -602.96418
#> 15:            1161.3052      3816186.5           1661077.26  -601.29338
#> 16:            1115.7519      4874383.8            602880.04   679.41846
#> 17:            1073.6378     30716146.3          25238882.48  -193.58740
#> 18:            1034.5877      3103595.8           2373668.03  -636.25035
#> 19:             998.2788      4615142.7            862121.11  -346.65873
#> 20:             964.4322      3599253.6           1878010.17  -417.02454
#> 21:             932.8057     10368256.1           4890992.31   571.33681
#> 22:             903.1878      2902259.5           2575004.30  -304.97706
#> 23:             875.3929      3752257.6           1725006.18  -421.44210
#> 24:             849.2579      2957872.0           2519391.84  -390.59687
#> 25:             824.6382      3439794.1           2037469.68  -574.03345
#> 26:             801.4058      3848104.2           1629159.62   -98.38990
#> 27:             779.4467      2674437.8           2802825.95  -321.34395
#> 28:             758.6590      4355955.3           1121308.46  -298.60772
#> 29:             738.9513      2894321.9           2582941.91  -438.92325
#> 30:             720.2417      3480731.8           1996532.04  -461.05871
#> 31:             702.4561      1316819.3           4160444.53    30.19482
#> 32:             685.5278      1825070.1           3652193.66  -658.97386
#> 33:             669.3962      2282415.9           3194847.88   -12.39375
#> 34:             654.0064      1859137.2           3618126.64  -543.90956
#> 35:             639.3084      1719022.4           3758241.38  -350.37996
#> 36:             625.2565      3940098.8           1537164.97   230.32393
#> 37:             611.8091      1735502.2           3741761.62  -342.68008
#> 38:             598.9279      2214468.8           3262795.01  -303.28410
#> 39:             586.5780      2857748.4           2619515.38  -291.38914
#> 40:             574.7271      2256914.1           3220349.71  -356.90423
#> 41:             563.3457     19415305.7          13938041.92  5938.25441
#> 42:             552.4062      3733155.8           1744107.99  -389.83595
#> 43:             541.8835      1756358.7           3720905.11  -119.10392
#> 44:             531.7543      2302478.4           3174785.37  -139.48348
#> 45:             521.9968      2927004.6           2550259.24  -245.91341
#> 46:             512.5909      3476476.8           2000786.96  -241.33694
#> 47:             503.5180      1907838.3           3569425.50  -127.70131
#> 48:             494.7607      1797239.2           3680024.59  -326.95742
#> 49:             486.3029      1327885.5           4149378.35  -231.05954
#> 50:             478.1293      8795981.3           3318717.45    38.03500
#> 51:             470.2260      7464316.2           1987052.43   303.36602
#> 52:             462.5797      2125295.3           3351968.50  -323.02921
#> 53:             455.1781      1197619.3           4279644.49  -105.11585
#> 54:             448.0097      2765697.2           2711566.65  -177.86121
#> 55:             441.0635      1727265.8           3749998.00    67.76534
#> 56:             434.3294      5680742.2            203478.37   -98.20743
#> 57:             427.7979      2331080.9           3146182.95  -264.75249
#> 58:             421.4599      1888640.1           3588623.66  -135.84438
#> 59:             415.3070      1301798.4           4175465.37  -254.45917
#> 60:             409.3312      1889586.2           3587677.63  -245.57428
#> 61:             403.5249      1920607.6           3556656.17   -58.16842
#> 62:             397.8810      2705223.8           2772039.95  -230.22148
#> 63:             392.3928      2715704.3           2761559.47  -120.71448
#> 64:             387.0539      1712036.2           3765227.62  -356.62946
#> 65:             381.8584      2302604.0           3174659.76   -88.94943
#> 66:             376.8006      4400397.5           1076866.33   239.34171
#> 67:             371.8749      2188426.7           3288837.07  -172.74079
#> 68:             367.0764      1999656.8           3477606.99  -274.18246
#> 69:             362.4001      2305037.6           3172226.24  -258.96959
#> 70:             357.8415      2091568.7           3385695.07  -191.08913
#> 71:             353.3962      3127479.0           2349784.82   147.47131
#> 72:             349.0599      1389333.5           4087930.31  -204.27649
#> 73:             344.8288      1475671.9           4001591.86   -71.32133
#> 74:             340.6990      2146793.7           3330470.13  -142.65253
#> 75:             336.6670      5446456.9             30806.95   -82.25993
#> 76:             332.7292      3105234.1           2372029.67    38.86618
#> 77:             328.8826      1636814.5           3840449.28   211.73502
#> 78:             325.1238      7035869.3           1558605.46  -214.64458
#> 79:             321.4500      1318511.8           4158752.02   -60.26419
#> 80:             317.8583      1680245.5           3797018.31   231.88423
#> 81:             314.3460      1689276.6           3787987.18   736.80836
#> 82:             310.9105      1247501.7           4229762.13   -28.48443
#> 83:             307.5492      3056765.5           2420498.32   709.01288
#> 84:             304.2599      1349480.5           4127783.33   224.30704
#> 85:             301.0401      2308419.1           3168844.74  -111.59551
#> 86:             297.8878      1565597.9           3911665.87   422.31141
#> 87:             294.8008      2101475.8           3375787.97   449.03373
#> 88:             291.7772       992128.7           4485135.10  -172.42287
#> 89:             288.8149      2658256.1           2819007.74   889.94705
#> 90:             285.9122      3009208.8           2468055.04   671.14872
#>     benford.so.dist.freq data.summation abs.excess.summation  difference
#>     squared.diff absolute.diff z.statistic
#>  1: 1.032084e+03    2811.91765  32.8065025
#>  2: 1.587369e+03    3331.97978  40.6105137
#>  3: 3.048523e+01     442.87491   5.6135423
#>  4: 5.986347e+00     188.83775   2.4804659
#>  5: 1.964185e+02    1043.68327  14.2229271
#>  6: 3.670452e+01     436.35975   6.1381165
#>  7: 4.305315e+01     458.03897   6.6423539
#>  8: 6.835990e+01     560.42335   8.3651065
#>  9: 5.456156e+01     486.95166   7.4671851
#> 10: 2.820015e+01     340.98196   5.3626610
#> 11: 6.431124e-01      50.22096   0.8025066
#> 12: 9.182811e+00     185.30331   3.0531379
#> 13: 7.598521e+01     521.05604   8.7938819
#> 14: 1.062758e+02     602.96418  10.3970023
#> 15: 1.101863e+02     601.29338  10.5824662
#> 16: 1.464227e+02     679.41846  12.1959339
#> 17: 1.235372e+01     193.58740   3.5347956
#> 18: 1.384810e+02     636.25035  11.8525190
#> 19: 4.260433e+01     346.65873   6.5680252
#> 20: 6.381941e+01     417.02454   8.0385232
#> 21: 1.238496e+02     571.33681  11.1990585
#> 22: 3.644665e+01     304.97706   6.0691953
#> 23: 7.180821e+01     421.44210   8.5210477
#> 24: 6.357981e+01     390.59687   8.0156185
#> 25: 1.414205e+02     574.03345  11.9571787
#> 26: 4.275135e+00      98.38990   2.0698331
#> 27: 4.688729e+01     321.34395   6.8778176
#> 28: 4.159650e+01     298.60772   6.4763485
#> 29: 9.227045e+01     438.92325   9.6493903
#> 30: 1.044565e+02     461.05871  10.2659081
#> 31: 4.593535e-01      30.19482   0.6701361
#> 32: 2.241881e+02     658.97386  15.0404619
#> 33: 8.121261e-02      12.39375   0.2748896
#> 34: 1.600931e+02     543.90956  12.7047423
#> 35: 6.796251e+01     350.37996   8.2726424
#> 36: 3.002764e+01     230.32393   5.4941368
#> 37: 6.793023e+01     342.68008   8.2686594
#> 38: 5.435331e+01     303.28410   7.3941970
#> 39: 5.122975e+01     291.38914   7.1774231
#> 40: 7.844097e+01     356.90423   8.8833400
#> 41: 2.215358e+04    5938.25441 149.4723916
#> 42: 9.736582e+01     389.83595   9.8965741
#> 43: 9.265044e+00     119.10392   3.0436909
#> 44: 1.294899e+01     139.48348   3.6002138
#> 45: 4.100133e+01     245.91341   6.4158233
#> 46: 4.021407e+01     241.33694   6.3532250
#> 47: 1.146244e+01     127.70131   3.3854027
#> 48: 7.646954e+01     326.95742   8.7644757
#> 49: 3.885458e+01     231.05954   6.2430733
#> 50: 1.070835e+00      38.03500   1.0249558
#> 51: 6.926734e+01     303.36602   8.3389680
#> 52: 7.983591e+01     323.02921   8.9529326
#> 53: 8.591249e+00     105.11585   2.9273296
#> 54: 2.499058e+01     177.86121   5.0021402
#> 55: 3.684812e+00      67.76534   1.9118706
#> 56: 7.859063e+00      98.20743   2.7984221
#> 57: 5.798861e+01     264.75249   7.6255856
#> 58: 1.549631e+01     135.84438   3.9347251
#> 59: 5.517829e+01     254.45917   7.4372271
#> 60: 5.214255e+01     245.57428   7.2288992
#> 61: 2.967601e+00      58.16842   1.7131498
#> 62: 4.714545e+01     230.22148   6.8722456
#> 63: 1.314314e+01     120.71448   3.6211927
#> 64: 1.162959e+02     356.62946  10.8008940
#> 65: 7.333060e+00      88.94943   2.7006236
#> 66: 5.380548e+01     239.34171   7.3410430
#> 67: 2.839842e+01     172.74079   5.3287395
#> 68: 7.248100e+01     274.18246   8.5219636
#> 69: 6.549535e+01     258.96959   8.0997315
#> 70: 3.611457e+01     191.08913   6.0102550
#> 71: 2.177983e+01     147.47131   4.6636612
#> 72: 4.230952e+01     204.27649   6.5060137
#> 73: 5.220795e+00      71.32133   2.2748832
#> 74: 2.113924e+01     142.65253   4.5935896
#> 75: 7.113403e+00      82.25993   2.6577237
#> 76: 1.606771e+00      38.86618   1.2544677
#> 77: 4.824428e+01     211.73502   6.9468694
#> 78: 5.015247e+01     214.64458   7.0829439
#> 79: 3.998586e+00      60.26419   1.9879396
#> 80: 5.987011e+01     231.88423   7.7396952
#> 81: 6.112268e+02     736.80836  24.7657203
#> 82: 9.235940e-01      28.48443   0.9464172
#> 83: 5.784887e+02     709.01288  24.0914542
#> 84: 5.852514e+01     224.30704   7.6509090
#> 85: 1.464099e+01     111.59551   3.8179938
#> 86: 2.118918e+02     422.31141  14.5724430
#> 87: 2.420642e+02     449.03373  15.5761772
#> 88: 3.606116e+01     172.42287   6.0010612
#> 89: 9.705323e+02     889.94705  31.2047330
#> 90: 5.575786e+02     671.14872  23.6471754
#>     squared.diff absolute.diff z.statistic