Shows the first digits ordered by the mains discrepancies from Benford's Law

It creates a data frame with the first digits and the differences from Benford's Law in decreasing order.

suspectsTable(bfd, by="absolute.diff")

Arguments

bfd
an object of class "Benford". See benford.

by
a character string selecting how to order the digits.It can be 'abs.excess.summation','difference','squared.diff' or 'absolute.diff'.

Value

A data frame with 2 variables: digits and the group chosen in by.

Examples

data(corporate.payment) #gets data 
cp <- benford(corporate.payment$Amount) #generates benford object
suspectsTable(cp)

#>     digits absolute.diff
#>  1:     50    5938.25441
#>  2:     11    3331.97978
#>  3:     10    2811.91765
#>  4:     14    1043.68327
#>  5:     98     889.94705
#>  6:     90     736.80836
#>  7:     92     709.01288
#>  8:     25     679.41846
#>  9:     99     671.14872
#> 10:     41     658.97386
#> 11:     27     636.25035
#> 12:     23     602.96418
#> 13:     24     601.29338
#> 14:     34     574.03345
#> 15:     30     571.33681
#> 16:     17     560.42335
#> 17:     43     543.90956
#> 18:     22     521.05604
#> 19:     18     486.95166
#> 20:     39     461.05871
#> 21:     16     458.03897
#> 22:     96     449.03373
#> 23:     12     442.87491
#> 24:     38     438.92325
#> 25:     15     436.35975
#> 26:     95     422.31141
#> 27:     32     421.44210
#> 28:     29     417.02454
#> 29:     33     390.59687
#> 30:     51     389.83595
#> 31:     49     356.90423
#> 32:     73     356.62946
#> 33:     44     350.37996
#> 34:     28     346.65873
#> 35:     46     342.68008
#> 36:     19     340.98196
#> 37:     57     326.95742
#> 38:     61     323.02921
#> 39:     36     321.34395
#> 40:     31     304.97706
#> 41:     60     303.36602
#> 42:     47     303.28410
#> 43:     37     298.60772
#> 44:     48     291.38914
#> 45:     77     274.18246
#> 46:     66     264.75249
#> 47:     78     258.96959
#> 48:     68     254.45917
#> 49:     54     245.91341
#> 50:     69     245.57428
#> 51:     55     241.33694
#> 52:     75     239.34171
#> 53:     89     231.88423
#> 54:     58     231.05954
#> 55:     45     230.32393
#> 56:     71     230.22148
#> 57:     93     224.30704
#> 58:     87     214.64458
#> 59:     86     211.73502
#> 60:     81     204.27649
#> 61:     26     193.58740
#> 62:     79     191.08913
#> 63:     13     188.83775
#> 64:     21     185.30331
#> 65:     63     177.86121
#> 66:     76     172.74079
#> 67:     97     172.42287
#> 68:     80     147.47131
#> 69:     83     142.65253
#> 70:     53     139.48348
#> 71:     67     135.84438
#> 72:     56     127.70131
#> 73:     72     120.71448
#> 74:     52     119.10392
#> 75:     94     111.59551
#> 76:     62     105.11585
#> 77:     35      98.38990
#> 78:     65      98.20743
#> 79:     74      88.94943
#> 80:     84      82.25993
#> 81:     82      71.32133
#> 82:     64      67.76534
#> 83:     88      60.26419
#> 84:     70      58.16842
#> 85:     20      50.22096
#> 86:     85      38.86618
#> 87:     59      38.03500
#> 88:     40      30.19482
#> 89:     91      28.48443
#> 90:     42      12.39375
#>     digits absolute.diff